+++ packages/vue-win95/src/components/date-picker/index.ts
@@ -5,4 +5,6 @@ export const W95DatePicker = withInstall(_DatePicker);
 export type W95DatePickerProps = {
   modelValue?: string | null;
   label?: string;
+  min?: string | null;
+  max?: string | null;
 };