+++ packages/vue-win95/src/composables/index.ts
@@ -1,4 +1,11 @@
 export { useId } from "./useId";
 export { useFocusTrap } from "./useFocusTrap";
 export { useRovingFocus } from "./useRovingFocus";
+export {
+  useNotifier,
+  notify,
+  notifyClear,
+  notifyItems,
+} from "./useNotifier";
 export type { Orientation, RovingFocusOptions } from "./useRovingFocus";
+export type { NotifyItem, NotifyType } from "./useNotifier";